[ICU-12736] assume in some places that we always have std::string; merge r39687 & r39704 from dev branch X-Commit-URL: https://ssl.icu-project.org/trac/changeset/39731 Former-commit-id: 66c36e6c0595073da4fcfc5b7f532fe0b2dbddf7

+++ icu4c/source/test/intltest/strcase.cpp
@@ -702,7 +702,8 @@ StringCaseTest::assertGreekUpper(const char *s, const char *expected) {
             expectedErrorCode = U_STRING_NOT_TERMINATED_WARNING;
         } else {
             expectedErrorCode = U_ZERO_ERROR;
-            assertEquals(msg + cap + " NUL", 0, dest8b[length]);
+            // Casts to int32_t to avoid matching UBool.
+            assertEquals(msg + cap + " NUL", (int32_t)0, (int32_t)dest8b[length]);
         }
         assertEquals(msg + cap + " errorCode", expectedErrorCode, errorCode);
         if (cap >= expected8Length) {
